Generate Random qoute from a file
This program reads from a text file called "quotes.txt" and stores each line of the file into an array called "quotes". The number of lines in the file is stored in the variable "line_count". The program then generates a random number between 0 and the number of lines in the file, and selects the quote at that index in the "quotes" array to be printed out.